+++ NEWS.md
@@ -8,6 +8,7 @@
 
 ## Bug fixes
 * Add catch in `runPCAprojectionBatch()` for when ncp requested exceeds number of feats used
+* Make `spatCellCellcom()` respect `verbose` flag [#949](https://github.com/drieslab/Giotto/issues/949) by rbutleriii
 
 
 # Giotto 4.0.8 (2024/05/22)